+++ src/Equinox.DynamoStore.Prometheus/DynamoStorePrometheus.fs
@@ -113,8 +113,10 @@ type LogSink(customTags: seq<string * string>) =
                 | Op       (Operation.Tip304,   m) -> observeTip  ("R", "query",    "tip",           "ok", "304") m
                 | Op       (Operation.Query,    m) -> observe     ("R", "query",    "query",         "ok")        m
                 | QueryRes (_direction,         m) -> observeRes  ("R", "query",    "queryPage")                  m
-                | Op       (Operation.Write,    m) -> observe     ("W", "transact", "sync",          "ok")        m
-                | Op       (Operation.Conflict, m) -> observe     ("W", "transact", "conflict",      "conflict")  m
+                | Op       (Operation.Append,   m) -> observe     ("W", "transact", "append",        "ok")        m
+                | Op       (Operation.Calve,    m) -> observe     ("W", "transact", "calve",         "ok")        m
+                | Op       (Operation.AppendConflict, m) ->observe("W", "transact", "append",         "conflict")  m
+                | Op       (Operation.CalveConflict, m) -> observe("W", "transact", "calve",         "conflict")  m
                 | Op       (Operation.Prune,    m) -> observe_    ("R", "prune",    "pruneQuery")                 m
                 | PruneRes                      m  -> observeRes  ("R", "prune",    "pruneQueryPage")             m
                 | Op       (Operation.Delete,   m) -> observe     ("W", "prune",    "delete",        "ok")        m