+++ tests/Equinox.CosmosStore.Integration/DocumentStoreIntegration.fs
@@ -105,7 +105,7 @@ type Tests(testOutputHelper) =
             do! addAndThenRemoveItemsManyTimesExceptTheLastOne cartContext cartId skuId service addRemoveCount
             test <@ i = i && List.replicate (expectedResponses (i-1)) EqxAct.ResponseBackward @ [EqxAct.QueryBackward; EqxAct.Append] = capture.ExternalCalls @>
 #if STORE_DYNAMO
-            if eventsInTip then verifyRequestChargesMax 181 // 180.5 [8.5; 172]
+            if eventsInTip then verifyRequestChargesMax 182 // 181.5 [5.5; 176]
 #else
             if eventsInTip then verifyRequestChargesMax 76 // 76.0 [3.72; 72.28]
 #endif